refactor: domain 패키지의 하위 패키지를 entity, vo로 분리
