+++ src/main/java/shopping/application/CartService.java
@@ -5,8 +5,8 @@ import org.springframework.transaction.annotation.Transactional;
 import shopping.domain.CartItems;
 import shopping.domain.entity.CartItem;
 import shopping.domain.entity.Product;
-import shopping.domain.entity.Quantity;
 import shopping.domain.entity.User;
+import shopping.domain.vo.Quantity;
 import shopping.dto.CartCreateRequest;
 import shopping.dto.CartResponse;
 import shopping.dto.QuantityUpdateRequest;
