+++ src/main/java/shopping/application/OrderService.java
@@ -3,10 +3,10 @@ package shopping.application;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import shopping.application.mapper.OrderMapper;
-import shopping.domain.ExchangeRate;
 import shopping.domain.entity.CartItem;
 import shopping.domain.entity.Order;
 import shopping.domain.entity.OrderItem;
+import shopping.domain.vo.ExchangeRate;
 import shopping.dto.OrderResponse;
 import shopping.exception.OrderNotFoundException;
 import shopping.repository.CartItemRepository;
