+++ src/main/java/shopping/domain/vo/Image.java
@@ -1,4 +1,4 @@
-package shopping.domain.entity;
+package shopping.domain.vo;
 
 import shopping.exception.ImageLengthInvalidException;
