+++ src/main/java/shopping/domain/vo/Quantity.java
@@ -1,4 +1,4 @@
-package shopping.domain.entity;
+package shopping.domain.vo;
 
 
 import shopping.exception.QuantityInvalidException;
