+++ src/main/java/shopping/ui/OrderController.java
@@ -4,7 +4,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import shopping.application.ExchangeRateService;
 import shopping.application.OrderService;
-import shopping.domain.ExchangeRate;
+import shopping.domain.vo.ExchangeRate;
 import shopping.dto.OrderResponse;
 import shopping.ui.config.AuthenticationPrincipal;
