+++ src/test/java/shopping/domain/CartItemsTest.java
@@ -4,9 +4,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import shopping.domain.entity.CartItem;
 import shopping.domain.entity.Product;
-import shopping.domain.entity.Quantity;
 import shopping.domain.entity.User;
-import shopping.domain.entity.fixture.EntityFixture;
+import shopping.domain.fixture.DomainFixture;
+import shopping.domain.vo.Quantity;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -19,8 +19,8 @@ class CartItemsTest {
     @DisplayName("장바구니에 상품을 추가한다.")
     void addCartItem() {
         // given
-        User user = EntityFixture.createUser();
-        Product product = EntityFixture.createProduct();
+        User user = DomainFixture.createUser();
+        Product product = DomainFixture.createProduct();
         CartItem item = new CartItem(user, product, Quantity.ONE);
         CartItems items = new CartItems(new ArrayList<>());
 
@@ -35,8 +35,8 @@ class CartItemsTest {
     @DisplayName("장바구니에 존재하는 상품을 추가하면 수량이 늘어난다.")
     void addCartItemDuplicate() {
         // given
-        User user = EntityFixture.createUser();
-        Product product = EntityFixture.createProduct();
+        User user = DomainFixture.createUser();
+        Product product = DomainFixture.createProduct();
         CartItem item = new CartItem(user, product, Quantity.ONE);
         CartItems items = new CartItems(new ArrayList<>());
 
@@ -53,8 +53,8 @@ class CartItemsTest {
     @DisplayName("장바구니에 상품이 존재하면 true를 반환한다.")
     void contains() {
         // given
-        User user = EntityFixture.createUser();
-        Product product = EntityFixture.createProduct();
+        User user = DomainFixture.createUser();
+        Product product = DomainFixture.createProduct();
         CartItem item = new CartItem(1L, user, product, Quantity.ONE);
         CartItems items = new CartItems(new ArrayList<>(List.of(item)));
 
@@ -67,8 +67,8 @@ class CartItemsTest {
     @DisplayName("장바구니에 상품이 존재하지 않으면 false를 반환한다.")
     void notContains() {
         // given
-        User user = EntityFixture.createUser();
-        Product product = EntityFixture.createProduct();
+        User user = DomainFixture.createUser();
+        Product product = DomainFixture.createProduct();
         CartItem item = new CartItem(1L, user, product, Quantity.ONE);
         CartItems items = new CartItems(new ArrayList<>());
