+++ src/test/java/shopping/domain/entity/CartItemTest.java
@@ -2,7 +2,8 @@ package shopping.domain.entity;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import shopping.domain.entity.fixture.EntityFixture;
+import shopping.domain.fixture.DomainFixture;
+import shopping.domain.vo.Quantity;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -12,8 +13,8 @@ class CartItemTest {
     @DisplayName("장바구니 항목 수량을 1 증가할 수 있다.")
     void increaseQuantity() {
         // given
-        User user = EntityFixture.createUser();
-        Product product = EntityFixture.createProduct();
+        User user = DomainFixture.createUser();
+        Product product = DomainFixture.createProduct();
         CartItem cartItem = new CartItem(user, product, Quantity.ONE);
 
         // when
@@ -27,8 +28,8 @@ class CartItemTest {
     @DisplayName("장바구니 항목 수량을 업데이트 할 수 있다.")
     void updateQuantity() {
         // given
-        User user = EntityFixture.createUser();
-        Product product = EntityFixture.createProduct();
+        User user = DomainFixture.createUser();
+        Product product = DomainFixture.createProduct();
         CartItem cartItem = new CartItem(user, product, Quantity.ONE);
 
         // when
