+++ src/test/java/shopping/domain/entity/UserTest.java
@@ -2,8 +2,8 @@ package shopping.domain.entity;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import shopping.domain.Email;
-import shopping.domain.Password;
+import shopping.domain.vo.Email;
+import shopping.domain.vo.Password;
 import shopping.infrastructure.SHA256PasswordEncoder;
 
 import static org.assertj.core.api.Assertions.assertThatNoException;
