+++ src/test/java/shopping/domain/vo/ExchangeRateTest.java
@@ -1,9 +1,8 @@
-package shopping.domain;
+package shopping.domain.vo;
 
 import org.assertj.core.data.Percentage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import shopping.domain.entity.Price;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
