+++ src/test/java/shopping/domain/vo/ImageTest.java
@@ -1,4 +1,4 @@
-package shopping.domain.entity;
+package shopping.domain.vo;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
