+++ src/test/java/shopping/dto/OrderResponseTest.java
@@ -3,7 +3,7 @@ package shopping.dto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import shopping.domain.entity.Order;
-import shopping.domain.entity.fixture.EntityFixture;
+import shopping.domain.fixture.DomainFixture;
 
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
@@ -13,7 +13,7 @@ class OrderResponseTest {
     @DisplayName("Order로 OrderResponse를 생성한다.")
     void createOrderResponse() {
         // given
-        final Order order = EntityFixture.createOrderWithId();
+        final Order order = DomainFixture.createOrderWithId();
 
         // when & then
         assertThatNoException().isThrownBy(() -> OrderResponse.from(order));
